gallery and slider

+++ app/Http/Controllers/AdminPanel/AdminTransferController.php
@@ -3,34 +3,23 @@
 namespace App\Http\Controllers\AdminPanel;
 
 use App\Http\Controllers\Controller;
+
 use App\Models\Category;
+use App\Models\Transfer;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Storage;
 
-class CategoryController extends Controller
+class AdminTransferController extends Controller
 {
-    protected $appends= [
-        'getParentsTree'
-    ];
-    public static function getParentsTree($category,$title)
-    {
-        if($category->parent_id == 0)
-        {
-            return $title;
 
-        }
-        $parent = Category::find($category->parent_id);
-      $title = $parent->title . ' > ' . $title;
-        return CategoryController::getParentsTree($parent,$title);
-    }
     /**
      * Display a listing of the resource.
      *
      * @return \Illuminate\Http\Response
      */
     public function index()
-    {   $data= Category::all();
-        return view('admin.category.index',[
+    {   $data= Transfer::all();
+        return view('admin.transfer.index',[
             'data'=>$data
         ]);
     }
@@ -42,7 +31,7 @@ class CategoryController extends Controller
      */
     public function create()
     {   $data= Category::all();
-        return view('admin.category.create',[
+        return view('admin.transfer.create',[
             'data'=>$data
         ]);
     }
@@ -55,29 +44,34 @@ class CategoryController extends Controller
      */
     public function store(Request $request)
     {
-        $data= new Category();
-        $data->parent_id= $request->parent_id;
+        $data= new Transfer();
+        $data->category_id= $request->category_id;
         $data->title=$request->title;
         $data->keywords=$request->keywords;
+        $data->detail=$request->detail;
+        $data->km_price=$request->km_price;
+        $data->base_price=$request->base_price;
+        $data->capasity=$request->capasity;
+        $data->type_car=$request->type_car;
         $data->description=$request->description;
         $data->status=$request->status;
         if($request->file('image')){
             $data->image= $request->file('image')->store('images');
         }
         $data->save();
-        return redirect('admin/category');
+        return redirect('admin/transfer');
     }
 
     /**
      * Display the specified resource.
      *
-     * @param  \App\Models\Category  $category
+     * @param  \App\Models\Transfer  $transfer
      * @return \Illuminate\Http\Response
      */
-    public function show(Category $category,$id )
+    public function show(Transfer $transfer,$id )
     {
-        {   $data= Category::find($id);
-            return view('admin.category.show',[
+        {   $data= Transfer::find($id);
+            return view('admin.transfer.show',[
                 'data'=>$data
             ]);
         }
@@ -86,14 +80,14 @@ class CategoryController extends Controller
     /**
      * Show the form for editing the specified resource.
      *
-     * @param  \App\Models\Category  $category
+     * @param  \App\Models\Transfer  $transfer
      * @return \Illuminate\Http\Response
      */
-    public function edit(Category $category,$id)
+    public function edit(Transfer $transfer,$id)
     {
-        {   $data= Category::find($id);
+        {   $data= Transfer::find($id);
             $datalist=  Category::all();
-            return view('admin.category.edit',[
+            return view('admin.transfer.edit',[
                 'data'=>$data,
                 'datalist'=>$datalist
             ]);
@@ -105,35 +99,43 @@ class CategoryController extends Controller
      * Update the specified resource in storage.
      *
      * @param  \Illuminate\Http\Request  $request
-     * @param  \App\Models\Category  $category
+     * @param  \App\Models\Transfer  $transfer
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, Category $category,$id)
+    public function update(Request $request, Transfer $transfer,$id)
     {
-        $data= Category::find($id);
-        $data->parent_id= $request->parent_id;
+        $data= Transfer::find($id);
+        $data->category_id= $request->category_id;
         $data->title=$request->title;
         $data->keywords=$request->keywords;
+        $data->detail=$request->detail;
+        $data->km_price=$request->km_price;
+        $data->base_price=$request->base_price;
+        $data->capasity=$request->capasity;
+        $data->type_car=$request->type_car;
         $data->description=$request->description;
         $data->status=$request->status;
         if($request->file('image')){
                $data->image= $request->file('image')->store('images');
         }
         $data->save();
-        return redirect('admin/category');
+        return redirect('admin/transfer');
+
     }
 
     /**
      * Remove the specified resource from storage.
      *
-     * @param  \App\Models\Category  $category
+     * @param  \App\Models\Transfer  $transfer
      * @return \Illuminate\Http\Response
      */
-    public function destroy(Category $category,$id)
+    public function destroy(Transfer $transfer,$id)
     {
-        $data= Category::find($id);
-        Storage::delete($data->image);
+        $data= Transfer::find($id);
+        if($data->image&& Storage::disk('public')->exists($data->image)){
+            Storage::delete($data->image);
+        }
         $data->delete();
-        return redirect('/admin/category');
+        return redirect('/admin/transfer');
     }
 }
